+++ arc/abc/__init__.py
@@ -2,7 +2,15 @@ from .client import Client
 from .command import CallableCommandBase, CallableCommandProto, CommandBase, CommandProto
 from .error_handler import HasErrorHandler
 from .hookable import Hookable, HookResult, with_hook, with_post_hook
-from .option import CommandOptionBase, Option, OptionBase, OptionParams, OptionWithChoices, OptionWithChoicesParams
+from .option import (
+    CommandOptionBase,
+    Option,
+    OptionBase,
+    OptionParams,
+    OptionType,
+    OptionWithChoices,
+    OptionWithChoicesParams,
+)
 from .plugin import PluginBase
 
 __all__ = (
@@ -13,6 +21,7 @@ __all__ = (
     "CallableCommandBase",
     "Option",
     "OptionBase",
+    "OptionType",
     "CommandOptionBase",
     "OptionParams",
     "OptionWithChoices",