+++ arc/command/option/int.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 import typing as t
 
 import attr
-import hikari
 
-from arc.abc.option import OptionWithChoices, OptionWithChoicesParams
+from arc.abc.option import OptionType, OptionWithChoices, OptionWithChoicesParams
 from arc.internal.types import ClientT
 
 if t.TYPE_CHECKING:
+    import hikari
     import typing_extensions as te
 
     from ...internal.types import AutocompleteCallbackT
@@ -97,8 +97,8 @@ class IntOption(OptionWithChoices[int, ClientT, IntParams[ClientT]]):
     """The maximum value of the option."""
 
     @property
-    def option_type(self) -> hikari.OptionType:
-        return hikari.OptionType.INTEGER
+    def option_type(self) -> OptionType:
+        return OptionType.INTEGER
 
     @classmethod
     def _from_params(cls, *, name: str, is_required: bool, params: IntParams[ClientT], **kwargs: t.Any) -> te.Self: