+++ arc/command/option/user.py
@@ -5,7 +5,7 @@ import typing as t
 import attr
 import hikari
 
-from arc.abc.option import CommandOptionBase, OptionParams
+from arc.abc.option import CommandOptionBase, OptionParams, OptionType
 from arc.internal.types import ClientT
 
 if t.TYPE_CHECKING:
@@ -47,8 +47,8 @@ class UserOption(CommandOptionBase[hikari.User, ClientT, UserParams]):
     """
 
     @property
-    def option_type(self) -> hikari.OptionType:
-        return hikari.OptionType.USER
+    def option_type(self) -> OptionType:
+        return OptionType.USER
 
     @classmethod
     def _from_params(cls, *, name: str, is_required: bool, params: UserParams, **kwargs: t.Any) -> te.Self: